refactor(web): add suspense
